:lipstick: notification style refactoring for desktop and mobile
