+++ src/screens/components/talk/deliberation/index.js
@@ -1,1 +1,1 @@
-export { default } from './deliberation.notification'
+export { default } from './deliberation'
