+++ src/screens/components/talk/deliberation/notification/index.js
@@ -1,0 +1,1 @@
+export { default } from './notification.container'
